refactor(web): login button 성공 이벤트 prop 제거
