+++ apps/web/src/features/login/ui/LoginBanner.tsx
@@ -1,23 +1,11 @@
 "use client";
 
 import { useLocale } from "@recap/i18n";
-import { useQueryClient } from "@recap/react-query";
 
-import { useAuth } from "@/entities/auth/ui";
 import LoginButton from "@/features/login/ui/LoginButton";
-import { USER_KEYS } from "@/features/settings/api/query-keys";
 
 const LoginBanner = () => {
   const { t } = useLocale("settings");
-  const { refreshAuth } = useAuth();
-  const queryClient = useQueryClient();
-
-  const handleLoginSuccess = async () => {
-    await queryClient.resetQueries({ queryKey: USER_KEYS.details() });
-    await queryClient.invalidateQueries({ queryKey: USER_KEYS.details() });
-
-    refreshAuth();
-  };
 
   return (
     <div className="bg-blue-75 rounded-[1.25rem] px-5 py-5 md:px-6 md:py-6 xl:px-9 xl:py-8">
@@ -26,10 +14,7 @@ const LoginBanner = () => {
           {t("account.dailyRecordsPrompt")}
         </h2>
 
-        <LoginButton
-          onLoginSuccess={handleLoginSuccess}
-          className="w-full justify-center md:w-auto md:justify-start"
-        />
+        <LoginButton className="w-full justify-center md:w-auto md:justify-start" />
       </div>
       <p className="text-body-2 md:text-heading-rg mt-2 text-gray-800">
         {t("account.sampleDataCaption")}
