+++ apps/web/src/features/login/ui/LoginButton.tsx
@@ -1,22 +1,29 @@
 "use client";
 
+import { useCallback } from "react";
 import { useLocale } from "@recap/i18n";
+import { useQueryClient } from "@recap/react-query";
 import { cn } from "@recap/ui";
 
+import { useAuth } from "@/entities/auth/ui";
 import { useGoogleTokenLogin } from "@/features/login/model/use-google-token-login";
+import { USER_KEYS } from "@/features/settings/api/query-keys";
 import RightIcon from "@/shared/assets/icons/arrow-right.svg";
 
-const LoginButton = ({
-  onLoginSuccess,
-  className,
-  children,
-}: {
-  onLoginSuccess?: () => void | Promise<void>;
-  className?: string;
-  children?: React.ReactNode;
-}) => {
-  const { ready, login } = useGoogleTokenLogin({ onLoginSuccess });
+const LoginButton = ({ className }: { className?: string }) => {
   const { t } = useLocale("settings");
+  const { refreshAuth } = useAuth();
+  const queryClient = useQueryClient();
+
+  const onLoginSuccess = useCallback(async () => {
+    await queryClient.resetQueries({ queryKey: USER_KEYS.details() });
+    await queryClient.invalidateQueries({ queryKey: USER_KEYS.details() });
+    refreshAuth();
+  }, [queryClient, refreshAuth]);
+
+  const { ready, login } = useGoogleTokenLogin({
+    onLoginSuccess,
+  });
 
   return (
     <button
@@ -28,7 +35,7 @@ const LoginButton = ({
         className,
       )}
     >
-      {children ?? t("account.login")}
+      {t("account.login")}
       <RightIcon />
     </button>
   );
